updated tests for the new hierarchy (#4464)
