+++ .gitignore
@@ -101,6 +101,7 @@ netdata-updater.sh
 
 # cmake files
 cmake-build-debug/
+cmake-build-release/
 CMakeCache.txt
 CMakeFiles/
 cmake_install.cmake
@@ -130,6 +131,7 @@ tests/profile/benchmark-registry
 tests/profile/test-eval
 tests/profile/benchmark-line-parsing
 tests/profile/benchmark-procfile-parser
+tests/profile/benchmark-value-pairs
 tests/profile/statsd-stress
 oprofile_data/
 vgcore.*
